+++ packages/filter-tag/src/component.screenshots.test.tsx
@@ -5,7 +5,7 @@ import {
     generateTestCases,
     customSnapshotIdentifier,
     createPreview,
-} from '../../screenshot-utils';
+} from '@alfalab/core-components-screenshot-utils';
 
 const screenshotTesting = setupScreenshotTesting({
     it,