+++ packages/switch/src/component.screenshots.test.tsx
@@ -4,7 +4,7 @@ import {
     createSpriteStorybookUrl,
     generateTestCases,
     createPreview,
-} from '../../screenshot-utils';
+} from '@alfalab/core-components-screenshot-utils';
 
 const screenshotTesting = setupScreenshotTesting({
     it,