+++ typings/react-textarea-autosize.d.ts
@@ -1,4 +1,0 @@
-declare module 'react-textarea-autosize' {
-    import TextareaAutosize from 'react-textarea-autosize/dist/declarations/src/index';
-    export default TextareaAutosize;
-}